last

+++ app/Http/Controllers/Backend/ProductVariantController.php
@@ -6,6 +6,7 @@ use App\DataTables\ProductVariantDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Product;
 use App\Models\ProductVariant;
+use App\Models\ProductVariantDetails;
 use Illuminate\Http\Request;
 
 class ProductVariantController extends Controller
@@ -84,6 +85,10 @@ class ProductVariantController extends Controller
     public function destroy(string $id)
     {
         $variant = ProductVariant::findOrFail($id);
+        $variantValuesCheckExist = ProductVariantDetails::where('product_variant_id', $variant->id)->count();
+        if ($variantValuesCheckExist > 0) {
+            return response(['status' => 'error', 'message' => 'this variant contains items inside, you must delete them first']);
+        }
         $variant->delete();
         return response(['status' => 'success', 'message' => 'Deleted successfully']);
     }
